+++ hbw/app/models/hbw/task.rb
@@ -45,8 +45,8 @@ module HBW
         with_user(email) do |user|
           do_request(:post,
                      'task/count',
-                     assignee:   user.id,
-                     active:     true)['count']
+                     assignee: user.id,
+                     active:   true)['count']
         end
       end
 
@@ -61,14 +61,15 @@ module HBW
 
       def fetch_unassigned(email, entity_class, first_result, max_results)
         entity_code_key = HBW::Widget.config[:entities].fetch(entity_class)[:entity_code_key]
+        bp_name_key = HBW::Widget.config[:entities].fetch(entity_class)[:bp_name_key]
 
         with_user(email) do |user|
           with_definitions(entity_code_key) do
             do_request(:post,
                        "task?firstResult=#{first_result}&maxResults=#{max_results}",
-                       active:         true,
-                       candidateUser:  user.id,
-                       sorting:        sorting_fields)
+                       active:        true,
+                       candidateUser: user.id,
+                       sorting:       sorting_fields(bp_name_key))
           end
         end
       end
@@ -95,7 +96,7 @@ module HBW
         end
       end
 
-      def sorting_fields
+      def sorting_fields(bp_name_key)
         [
           {
             sortBy:    'dueDate',
@@ -106,8 +107,12 @@ module HBW
             sortOrder: 'desc'
           },
           {
-            sortBy:    'name',
-            sortOrder: 'asc'
+            sortBy:    'processVariable',
+            sortOrder: 'asc',
+            parameters: {
+              variable: bp_name_key,
+              type:     'String'
+            }
           },
           {
             sortBy:    'created',